cast auf long explizit ausgeführt anstatt implizit zur Laufzeit

+++ app/fhemobile/src/main/java/de/fhe/fhemobile/canteencardbalance/canteencardreader/desfire/util/DesFireUtils.java
@@ -52,7 +52,7 @@ public class DesFireUtils {
         long value = 0;
         for (int i = 0; i < length; i++) {
             int shift = (length - 1 - i) * 8;
-            value += (b[i + offset] & 0x000000FF) << shift;
+            value += (long) (b[i + offset] & 0x000000FF) << shift;
         }
         return value;
     }
